+++ lib/silgy_lib.h
@@ -348,7 +348,7 @@ extern "C" {
     int getpid(void);
     int clock_gettime(int dummy, struct timespec *spec);
     char *stpcpy(char *dest, const char *src);
-    char *stpncpy(char *dest, const char *src, unsigned long len);
+    char *stpncpy(char *dest, const char *src, unsigned int len);
 #endif  /* _WIN32 */
 
 #ifndef strnstr
